+++ package.json
@@ -19,6 +19,11 @@
         "@nomiclabs/hardhat-ethers": "^2.1.1",
         "@nomiclabs/hardhat-etherscan": "^3.1.0",
         "@nomiclabs/hardhat-waffle": "^2.0.3",
+        "@typechain/ethers-v5": "^10.1.0",
+        "@typechain/hardhat": "^6.1.3",
+        "@types/chai": "^4.3.1",
+        "@types/mocha": "^9.1.1",
+        "@types/node": "^12.20.55",
         "chai": "^4.3.6",
         "dotenv": "^16.0.2",
         "esm": "3.2.25",
@@ -26,6 +31,7 @@
         "ethers": "^5.7.1",
         "glob": "^8.0.3",
         "hardhat": "^2.11.2",
+        "hardhat-contract-sizer": "^2.6.1",
         "hardhat-deploy": "^0.11.15",
         "hardhat-deploy-ethers": "^0.3.0-beta.13",
         "hardhat-gas-reporter": "^1.0.9",
@@ -34,6 +40,9 @@
         "solhint": "^3.3.7",
         "solhint-plugin-prettier": "^0.0.5",
         "solidity-coverage": "^0.8.2",
+        "solidity-docgen": "^0.6.0-beta.25",
+        "ts-node": "^10.8.1",
+        "typechain": "^8.1.0",
         "typescript": "^4.8.3"
     },
     "scripts": {
