+++ tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "compilerOptions": {
+        "target": "es2018",
+        "module": "commonjs",
+        "strict": false,
+        "esModuleInterop": true,
+        "resolveJsonModule": true,
+        "outDir": "dist",
+        "declaration": true,
+        "noImplicitAny": false
+    },
+    "include": [
+        "./scripts",
+        "./test",
+        "./typechain-types"
+    ],
+    "exclude": [
+        "node_modules"
+    ],
+    "files": [
+        "./hardhat.config.ts"
+    ]
+}
